Notifies are saved in database if price and size values match with search filter in profile, this happens when a new advert is created

+++ Skeleton/src/main/java/org/sample/controller/UserController.java
@@ -71,7 +71,7 @@ public class UserController {
      * @param value The id of the search that has to be deleted from the user profile.
      */
     @RequestMapping(value = "/saved-searches")
-    public ModelAndView showSearches(@RequestParam(required = false) Long value, Integer filter) {
+    public ModelAndView showSearches(@RequestParam(required = false) Long value, Long filter) {
     	System.out.println(value);
 	    ModelAndView model = new ModelAndView("profilesearches");
